Introuce `nat64` module
